refactor: access to a mock database stored in an adapter. Job-specific auxiliary functions placed in an extra package utils.
