+++ services/job/core/job_test.go
@@ -2,73 +2,17 @@ package core_test
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/google/uuid"
+	repo_in_memory "github.com/informatik-mannheim/cmg-ss2025/services/job/adapters/repo-in-memory"
 	"github.com/informatik-mannheim/cmg-ss2025/services/job/core"
 	"github.com/informatik-mannheim/cmg-ss2025/services/job/ports"
 )
 
-// MockJobStorage is a mock implementation of the JobStorage interface
-type MockJobStorage struct {
-	jobs map[string]ports.Job
-}
-
-func NewMockJobStorage() *MockJobStorage {
-	return &MockJobStorage{
-		jobs: make(map[string]ports.Job),
-	}
-}
-
-func (m *MockJobStorage) GetJobs(ctx context.Context, status []ports.JobStatus) ([]ports.Job, error) {
-	var results []ports.Job
-	for _, job := range m.jobs {
-		if len(status) == 0 || containsStatus(status, job.Status) {
-			results = append(results, job)
-		}
-	}
-	return results, nil
-}
-
-func (m *MockJobStorage) CreateJob(ctx context.Context, job ports.Job) error {
-	m.jobs[job.Id] = job
-	return nil
-}
-
-func (m *MockJobStorage) GetJob(ctx context.Context, id string) (ports.Job, error) {
-	job, ok := m.jobs[id]
-	if !ok {
-		return ports.Job{}, errors.New("job not found")
-	}
-	return job, nil
-}
-
-func (m *MockJobStorage) UpdateJob(ctx context.Context, id string, updateFn func(*ports.Job) error) (ports.Job, error) {
-	job, ok := m.jobs[id]
-	if !ok {
-		return ports.Job{}, errors.New("job not found")
-	}
-
-	if err := updateFn(&job); err != nil {
-		return ports.Job{}, err
-	}
-	m.jobs[id] = job
-	return job, nil
-}
-
-func containsStatus(statusList []ports.JobStatus, status ports.JobStatus) bool {
-	for _, s := range statusList {
-		if s == status {
-			return true
-		}
-	}
-	return false
-}
-
 func setup() *core.JobService {
-	storage := NewMockJobStorage()
+	storage := repo_in_memory.NewMockJobStorage()
 	return core.NewJobService(storage)
 }
 
@@ -319,7 +263,7 @@ func TestJobService_GetJobs(t *testing.T) {
 			wantErr: false,
 			setup: func() *core.JobService {
 				// Create a new service with an empty storage for this case
-				return core.NewJobService(NewMockJobStorage())
+				return core.NewJobService(repo_in_memory.NewMockJobStorage())
 			},
 		},
 		{
